chore: break out project assembly tests from project module
